upgrade C API support to 2.5.0 version

+++ pd4j_engine/src/main/java/com/am1goo/playdate4j/sdk/Sound.java
@@ -1889,7 +1889,7 @@ public class Sound {
 			bridge.soundSequence().setLoops(ptr.getValue(), startStep, endStep, loops);
 		}
 		
-		public int getTempo() {
+		public float getTempo() {
 			return bridge.soundSequence().getTempo(ptr.getValue());
 		}
